requirements.txt
flask==3.0.2
python-dotenv==1.0.0
exifread==3.0.0
google-cloud-vision==3.4.1
requests==2.31.0
werkzeug==3.0.1